+++ src/components/router.tsx
@@ -2,9 +2,10 @@ import Home from "@/pages/Home";
 import { Routes, Route } from "react-router-dom";
 import Layout from "./Layout";
 import SignIn from "@/pages/SignIn";
-import Products from "@/pages/Products";
-import ProductDetail from "@/pages/ProductDetail";
-import Cart from "@/pages/Cart";
+import About from "@/pages/About";
+import Products from "@/pages/products/Products";
+import ProductDetail from "@/pages/products/ProductDetail";
+import Cart from "@/pages/products/Cart";
 
 export default function Router() {
   return (
@@ -15,6 +16,7 @@ export default function Router() {
         <Route path="/products" element={<Products />} />
         <Route path="/products/:id" element={<ProductDetail />} />
         <Route path="/cart" element={<Cart />} />
+        <Route path="/about" element={<About />} />
       </Route>
     </Routes>
   );
