+++ src/pages/products/Cart.tsx
@@ -1,5 +1,6 @@
 import { CartItem } from "@/components/cart/CartItem";
 import BuyBtn from "@/components/common/BuyBtn";
+import EmptyProducts from "@/components/common/EmptyProducts";
 import { cartSelector } from "@/store";
 import { useRecoilState } from "recoil";
 
@@ -12,11 +13,15 @@ export default function Cart() {
 
   return (
     <div className="flex flex-col border md:p-3 gap-y-10 md:flex-row md:items-start md:justify-between">
-      <ul className="flex flex-col gap-y-3">
-        {cartList.map((item) => (
-          <CartItem key={item.id} {...item} setCartList={setCartList} />
-        ))}
-      </ul>
+      {cartList.length > 0 ? (
+        <ul className="flex flex-col gap-y-3">
+          {cartList.map((item) => (
+            <CartItem key={item.id} {...item} setCartList={setCartList} />
+          ))}
+        </ul>
+      ) : (
+        <EmptyProducts />
+      )}
 
       <div className="flex items-center justify-center p-3 gap-x-5 md:flex-col md:gap-y-5">
         <h1 className="text-xl">총 가격 : $ {totalPrice}</h1>
